[0.1.4] GameEngine is mostly done.
A few UI methods need to be added or slightly changed based on what I
realized we need. Basically just waiting on the UI for now to finish
things up. I'll probably clean up the code a bit later. 
In the Player class, the get and set life methods were private. I
changed them to public.